fixed images, double warning prevention.

+++ Events/Client/imageChecker.js
@@ -9,7 +9,7 @@ module.exports = {
    * @param {Client} client
    */
   async execute(message, client) {
-    const guild = client.guilds.cache.get(message.guild.id);
+    const guild = client.guilds.cache.get("946518364216520774");
     if (message.toString().startsWith("https")) {
       axios
         .get("https://api.moderatecontent.com/moderate/", {
@@ -68,7 +68,7 @@ Teen rating: \`${response.data.predictions.teen}\`
     } else if (message.attachments.size < 1) {
       return;
     } else {
-      message.attachments.forEach(async (m) => {
+      message.attachments.every((m) => {
         if (m.url.endsWith(".mov")) return;
         if (m.url.endsWith(".mp4")) return;
         axios
@@ -81,7 +81,7 @@ Teen rating: \`${response.data.predictions.teen}\`
           .then((response) => {
             if (response.data.predictions.adult > 75) {
               const channel = guild.channels.cache.get(message.channel.id);
-              try {
+              if (message.attachments.size > 1) {
                 message.delete();
                 channel.send({
                   content: `${message.author}`,
@@ -93,9 +93,36 @@ Teen rating: \`${response.data.predictions.teen}\`
                       .setColor("BLURPLE"),
                   ],
                 });
-              } catch (e) {
-                console.log(e.toString());
+                const logs = guild.channels.cache.get("1028748150149763092");
+                logs.send({
+                  embeds: [
+                    new MessageEmbed()
+                      .setDescription(
+                        `
+  ${message.author} (${message.author.id}) sent an image in ${message.channel}.
+  
+  Adult rating: \`${response.data.predictions.adult}\`
+  Everyone rating: \`${response.data.predictions.everyone}\`
+  Teen rating: \`${response.data.predictions.teen}\`
+              `
+                      )
+                      .setColor("BLURPLE")
+                      .setImage(`${m.url}`),
+                  ],
+                });
+                return;
               }
+              message.delete();
+              channel.send({
+                content: `${message.author}`,
+                embeds: [
+                  new MessageEmbed()
+                    .setDescription(
+                      "Your message has been deleted due to containing sensitive content."
+                    )
+                    .setColor("BLURPLE"),
+                ],
+              });
             }
             const logs = guild.channels.cache.get("1028748150149763092");
             logs.send({
